backend/requirements.txt
Flask==2.3.2
Werkzeug==2.3.2
firebase-admin==6.0.1
numpy==1.25.2
faiss-cpu==1.7.4
tensorflow==2.14.0